+++ internal/cli/plan.go
@@ -58,7 +58,7 @@ func (c *PlanCommand) Run(args []string) int {
 	packManager := generatePackManager(c.baseCommand, client, c.packConfig)
 
 	// load pack
-	r, err := renderPack(packManager, c.baseCommand.ui, errorContext)
+	r, err := renderPack(packManager, c.baseCommand.ui, false, errorContext)
 	if err != nil {
 		return c.exitCodeError
 	}