+++ internal/cli/render.go
@@ -23,12 +23,19 @@ import (
 type RenderCommand struct {
 	*baseCommand
 	packConfig *cache.PackConfig
+
 	// renderOutputTemplate is a boolean flag to control whether the output
 	// template is rendered.
 	renderOutputTemplate bool
+
 	// renderToDir is the path to write rendered job files to in addition to
 	// standard output.
 	renderToDir string
+
+	// renderAuxFiles is a boolean flag to control whether we should also render
+	// auxiliary files inside templates/
+	renderAuxFiles bool
+
 	// overwriteAll is set to true when someone specifies "a" to the y/n/a
 	overwriteAll bool
 }
@@ -190,7 +197,7 @@ func (c *RenderCommand) Run(args []string) int {
 	}
 	packManager := generatePackManager(c.baseCommand, client, c.packConfig)
 
-	renderOutput, err := renderPack(packManager, c.baseCommand.ui, errorContext)
+	renderOutput, err := renderPack(packManager, c.baseCommand.ui, c.renderAuxFiles, errorContext)
 	if err != nil {
 		return 1
 	}
@@ -202,7 +209,7 @@ func (c *RenderCommand) Run(args []string) int {
 		return 1
 	}
 
-	var renders = []Render{}
+	var renders []Render
 
 	// Iterate the rendered files and add these to the list of renders to
 	// output. This allows errors to surface and end things without emitting
@@ -280,6 +287,14 @@ func (c *RenderCommand) Flags() *flag.Sets {
 			Default: false,
 			Usage: `Controls whether or not the output template file within the
 					pack is rendered and displayed.`,
+		})
+
+		f.BoolVar(&flag.BoolVar{
+			Name:    "render-aux-files",
+			Target:  &c.renderAuxFiles,
+			Default: true,
+			Usage: `Controls whether or not the rendered output contains auxiliary
+					files found in the 'templates' folder.`,
 		})
 
 		f.StringVarP(&flag.StringVarP{