+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 IMPROVEMENTS:
+* cli: Add flag `--render-aux-files` that allows to render any files found in `templates/` [[GH-303](https://github.com/hashicorp/nomad-pack/pull/303)]
 * deps: Update the Nomad OpenAPI depedency [[GH-288](https://github.com/hashicorp/nomad-pack/pull/288)] and require Go 1.18 as a build dependency
 
 ## 0.0.1-techpreview.3 (July 21, 2022)